+++ src/quandary/ui/widgets/DebugTab.py
@@ -4,10 +4,6 @@ from quandary.ui.widgets.DebugPanel import DebugPanel
 
 class DebugTab(Static):
     """the debug tab of the main navigation window"""
-
-    DEFAULT_CSS = """
-    DebugTab {}
-    """
     
     def compose(self) -> ComposeResult:
         yield DebugPanel()
