+++ src/quandary/ui/widgets/ResponseTab.py
@@ -1,13 +1,13 @@
 from textual.widgets import Static
 from textual.app import ComposeResult
+from textual.containers import Vertical
 from quandary.ui.widgets.ResponsePanel import ResponsePanel
+from quandary.ui.widgets.InputPanel import InputPanel
 
 class ResponseTab(Static):
     """the response tab of the main navigation window"""
 
-    DEFAULT_CSS = """
-    ResponseTab {}
-    """
-
     def compose(self) -> ComposeResult:
-        yield ResponsePanel()
+        with Vertical():
+            yield ResponsePanel()
+            yield InputPanel()
