+++ src/quandary/ui/widgets/SettingsTab.py
@@ -6,14 +6,6 @@ from quandary.ui.widgets.SwitchSetting import SwitchSetting
 class SettingsTab(Static):
     """the settings tab of the main navigation window."""
 
-    DEFAULT_CSS = """
-    SettingsTab {
-        width: 1fr;
-        height: auto;
-        border: heavy $primary;
-    }
-    """
-
     def compose(self) -> ComposeResult:
         yield Horizontal(
             SwitchSetting(id="setting_doc_mode", enabled=True, label="Document Mode:      "),
